orquestador/pyproject.toml
[project]
name = "orchestrator"
version = "1.0.0"
description = "LangChain-based A2A Orchestrator Agent that coordinates multiple specialized agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.2",
    "click>=8.2.1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[project.scripts]
orchestrator = "app.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
